[CORRECTIVE] Fixed validation of long integers.
git-svn-id: svn://svn.code.sf.net/p/kactus2/code/branches/std2014_update@2325 f0aa6db5-7f46-43cc-a8a5-a31efdaafa67

+++ IPXACTmodels/Component/validators/AddressBlockValidator.h
@@ -204,7 +204,7 @@ private:
      *
      *      @return The register size in LAU.
      */
-    int getRegisterSizeInLAU(QSharedPointer<Register> targetRegister, int addressUnitBits) const;
+    qint64 getRegisterSizeInLAU(QSharedPointer<Register> targetRegister, int addressUnitBits) const;
 
     /*!
      *  Find errors within a name.
